+++ dpws/src/main/java/org/somda/sdc/dpws/http/apache/CommunicationLogHttpRequestInterceptor.java
@@ -7,8 +7,8 @@ import org.apache.http.HttpRequest;
 import org.apache.http.HttpRequestInterceptor;
 import org.apache.http.protocol.HttpContext;
 import org.apache.http.protocol.HttpCoreContext;
-import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.somda.sdc.common.logging.InstanceLogger;
 import org.somda.sdc.dpws.CommunicationLog;
 import org.somda.sdc.dpws.http.apache.helper.ApacheClientHelper;
@@ -19,12 +19,16 @@ import org.somda.sdc.dpws.soap.TransportInfo;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.util.Collections;
+import java.util.UUID;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Request interceptor which writes the outgoing request message and headers into the {@linkplain CommunicationLog}.
  */
 public class CommunicationLogHttpRequestInterceptor implements HttpRequestInterceptor {
     private static final Logger LOG = LogManager.getLogger(CommunicationLogHttpRequestInterceptor.class);
+    private static final String TRANSACTION_ID_PREFIX_CLIENT = "rrId:client:" + UUID.randomUUID() + ":";
+    private final static AtomicLong TRANSACTION_ID = new AtomicLong(-1L);
 
     private final CommunicationLog commlog;
     private final Logger instanceLogger;
@@ -41,8 +45,12 @@ public class CommunicationLogHttpRequestInterceptor implements HttpRequestInterc
         HttpHost target = (HttpHost) context.getAttribute(
                 HttpCoreContext.HTTP_TARGET_HOST);
 
+        var currentTransactionId = TRANSACTION_ID_PREFIX_CLIENT + TRANSACTION_ID.incrementAndGet();
+        context.setAttribute(CommunicationLog.MessageType.REQUEST.name(), currentTransactionId);
+
         var requestHttpApplicationInfo = new HttpApplicationInfo(
-                ApacheClientHelper.allHeadersToMultimap(request.getAllHeaders())
+                ApacheClientHelper.allHeadersToMultimap(request.getAllHeaders()),
+                currentTransactionId
         );
 
         // collect information for TransportInfo
@@ -60,6 +68,7 @@ public class CommunicationLogHttpRequestInterceptor implements HttpRequestInterc
         OutputStream commlogStream = commlog.logMessage(
                 CommunicationLog.Direction.OUTBOUND,
                 CommunicationLog.TransportType.HTTP,
+                CommunicationLog.MessageType.REQUEST,
                 requestCommContext);
 
         if (!(request instanceof HttpEntityEnclosingRequest)) {