+++ dpws/src/main/java/org/somda/sdc/dpws/http/jetty/CommunicationLogHandlerWrapper.java
@@ -13,11 +13,15 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.util.UUID;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * {@linkplain HandlerWrapper} which enables {@linkplain CommunicationLog} capabilities for requests and responses.
  */
 public class CommunicationLogHandlerWrapper extends HandlerWrapper {
+    private static final String TRANSACTION_ID_PREFIX_SERVER = "rrId:server:" + UUID.randomUUID() + ":";
+    private static final AtomicLong TRANSACTION_ID = new AtomicLong(-1L);
 
     private final CommunicationLog commLog;
     private final String frameworkIdentifier;
@@ -31,8 +35,12 @@ public class CommunicationLogHandlerWrapper extends HandlerWrapper {
     public void handle(String target, Request baseRequest, HttpServletRequest request, HttpServletResponse response)
             throws IOException, ServletException {
 
+        var currentTransactionId = TRANSACTION_ID_PREFIX_SERVER + TRANSACTION_ID.incrementAndGet();
+        baseRequest.setAttribute(CommunicationLog.MessageType.REQUEST.name(), currentTransactionId);
+
         var requestHttpApplicationInfo = new HttpApplicationInfo(
-                JettyUtil.getRequestHeaders(request)
+                JettyUtil.getRequestHeaders(request),
+                currentTransactionId
         );
 
         // collect information for TransportInfo
@@ -51,6 +59,7 @@ public class CommunicationLogHandlerWrapper extends HandlerWrapper {
         OutputStream input = commLog.logMessage(
                 CommunicationLog.Direction.INBOUND,
                 CommunicationLog.TransportType.HTTP,
+                CommunicationLog.MessageType.REQUEST,
                 requestCommContext);
         var out = baseRequest.getResponse().getHttpOutput();
 
@@ -65,7 +74,8 @@ public class CommunicationLogHandlerWrapper extends HandlerWrapper {
                     previousInterceptor,
                     commLog,
                     transportInfo,
-                    frameworkIdentifier
+                    frameworkIdentifier,
+                    currentTransactionId
             );
             out.setInterceptor(outInterceptor);