+++ dpws/src/test/java/it/org/somda/sdc/dpws/CommunicationLogIT.java
@@ -2,17 +2,18 @@ package it.org.somda.sdc.dpws;
 
 import com.google.common.collect.ListMultimap;
 import com.google.inject.AbstractModule;
+import com.google.inject.Injector;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.entity.ByteArrayEntity;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.util.EntityUtils;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.LogManager;
 import org.somda.sdc.dpws.CommunicationLog;
 import org.somda.sdc.dpws.CommunicationLogImpl;
 import org.somda.sdc.dpws.CommunicationLogSink;
@@ -42,6 +43,7 @@ import java.io.OutputStream;
 import java.net.InetSocketAddress;
 import java.net.URI;
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicReference;
@@ -80,6 +82,7 @@ public class CommunicationLogIT extends DpwsTest {
         this.overrideBindings(List.of(dpwsOverride, override));
         super.setUp();
 
+
         httpServerRegistry = getInjector().getInstance(JettyHttpServerRegistry.class);
         transportBindingFactory = getInjector().getInstance(TransportBindingFactory.class);
         soapMessageFactory = getInjector().getInstance(SoapMessageFactory.class);
@@ -95,7 +98,6 @@ public class CommunicationLogIT extends DpwsTest {
         marshalling.stopAsync().awaitTerminated();
         logSink.clear();
     }
-
 
     @Test
     void testClientCommlog() throws Exception {
@@ -134,7 +136,9 @@ public class CommunicationLogIT extends DpwsTest {
 
         // make requests to our server
         TransportBinding httpBinding1 = transportBindingFactory.createHttpBinding(baseUri.toString());
+
         for (int i = 0; i < 100; i++) {
+
             var requestMessage = createASoapMessage();
 
             ByteArrayOutputStream actualRequestStream = new ByteArrayOutputStream();
@@ -179,6 +183,11 @@ public class CommunicationLogIT extends DpwsTest {
                             .get(HttpServerUtil.GzipResponseHandler.TEST_HEADER_KEY)
                             .contains(HttpServerUtil.GzipResponseHandler.TEST_HEADER_VALUE)
             );
+
+            assertEquals(CommunicationLog.MessageType.RESPONSE, logSink.getInboundMessageType());
+            assertEquals(CommunicationLog.MessageType.REQUEST, logSink.getOutboundMessageType());
+
+            compareTransactionIds(logSink.getInboundTransactionIds(), logSink.getOutboundTransactionIds());
 
             logSink.clear();
         }
@@ -261,6 +270,12 @@ public class CommunicationLogIT extends DpwsTest {
                             .get(JettyHttpServerHandler.SERVER_HEADER_KEY)
                             .contains(JettyHttpServerHandler.SERVER_HEADER_VALUE)
             );
+
+            assertEquals(CommunicationLog.MessageType.REQUEST, logSink.getInboundMessageType());
+            assertEquals(CommunicationLog.MessageType.RESPONSE, logSink.getOutboundMessageType());
+
+            compareTransactionIds(logSink.getInboundTransactionIds(), logSink.getOutboundTransactionIds());
+
             logSink.clear();
         }
     }
@@ -343,6 +358,145 @@ public class CommunicationLogIT extends DpwsTest {
         }
     }
 
+    @Test
+    void testDuplicateTransactionIds() throws Exception {
+
+        final Injector secondInjector = configureInjector(List.of(
+                new DefaultDpwsConfigModule() {
+                    @Override
+                    public void customConfigure() {
+                        // ensure commlog works with compression enabled and doesn't store compressed messages
+                        bind(DpwsConfig.HTTP_GZIP_COMPRESSION, Boolean.class, true);
+                    }
+                },
+                new AbstractModule() {
+                    @Override
+                    protected void configure() {
+                        bind(CommunicationLogSink.class).toInstance(logSink);
+                        bind(CommunicationLog.class).to(CommunicationLogImpl.class).asEagerSingleton();
+                    }
+                }
+        ));
+
+        final JettyHttpServerRegistry secondHttpServerRegistry = secondInjector.getInstance(JettyHttpServerRegistry.class);
+        final TransportBindingFactory secondTransportBindingFactory = secondInjector.getInstance(TransportBindingFactory.class);
+        final SoapMessageFactory secondSoapMessageFactory = secondInjector.getInstance(SoapMessageFactory.class);
+        final EnvelopeFactory secondEnvelopeFactory = secondInjector.getInstance(EnvelopeFactory.class);
+        secondInjector.getInstance(JaxbMarshalling.class).startAsync().awaitRunning();
+        final TestCommLogSink secondLogSink = (TestCommLogSink) secondInjector.getInstance(CommunicationLogSink.class);
+        final SoapMarshalling secondMarshalling = secondInjector.getInstance(SoapMarshalling.class);
+        secondMarshalling.startAsync().awaitRunning();
+
+
+        URI baseUri = URI.create("http://127.0.0.1:0/");
+        String expectedResponse = "Sehr geehrter Kaliba, netter Versuch\n" +
+                "Kritische Texte, Weltverbesserer-Blues;";
+
+        JAXBElement<String> jaxbElement = new JAXBElement<>(
+                new QName("root-element"),
+                String.class, expectedResponse
+        );
+
+        var responseEnvelope = createASoapMessage();
+        responseEnvelope.getOriginalEnvelope().getBody().getAny().add(jaxbElement);
+
+        // make bytes out of the expected response
+        ByteArrayOutputStream expectedResponseStream = new ByteArrayOutputStream();
+        marshalling.marshal(responseEnvelope.getEnvelopeWithMappedHeaders(), expectedResponseStream);
+
+        var responseBytes = expectedResponseStream.toByteArray();
+
+        // spawn the http server
+        var handler = new HttpServerUtil.GzipResponseHandler(responseBytes);
+        var inetSocketAddress = new InetSocketAddress(baseUri.getHost(), baseUri.getPort());
+        var server = HttpServerUtil.spawnHttpServer(inetSocketAddress, handler);
+
+        // replace the port
+        baseUri = new URI(
+                baseUri.getScheme(),
+                baseUri.getUserInfo(),
+                baseUri.getHost(),
+                server.getAddress().getPort(),
+                baseUri.getPath(),
+                baseUri.getQuery(),
+                baseUri.getFragment());
+
+        // make requests to our server
+        TransportBinding httpBinding1 = transportBindingFactory.createHttpBinding(baseUri.toString());
+
+        TransportBinding httpBinding2 = secondTransportBindingFactory.createHttpBinding(baseUri.toString());
+        for (int i = 0; i < 100; i++) {
+
+            testSharedLogSink(marshalling, httpBinding1, logSink, expectedResponseStream);
+            testSharedLogSink(secondMarshalling, httpBinding2, secondLogSink, expectedResponseStream);
+
+            assertEquals(CommunicationLog.MessageType.RESPONSE, logSink.getInboundMessageType());
+            assertEquals(CommunicationLog.MessageType.REQUEST, logSink.getOutboundMessageType());
+
+            compareTransactionIds(logSink.getInboundTransactionIds(), logSink.getOutboundTransactionIds());
+
+            logSink.clear();
+        }
+    }
+
+    private void testSharedLogSink(SoapMarshalling soapMarshalling, TransportBinding httpBinding, TestCommLogSink logSink, ByteArrayOutputStream expectedResponseStream) throws Exception {
+        var requestMessage2 = createASoapMessage();
+
+        ByteArrayOutputStream actualRequestStream2 = new ByteArrayOutputStream();
+        soapMarshalling.marshal(requestMessage2.getEnvelopeWithMappedHeaders(), actualRequestStream2);
+
+        SoapMessage response2 = httpBinding.onRequestResponse(requestMessage2);
+
+        ByteArrayOutputStream actualResponseStream2 = new ByteArrayOutputStream();
+        soapMarshalling.marshal(response2.getEnvelopeWithMappedHeaders(), actualResponseStream2);
+
+        // response bytes should exactly match our expected bytes
+        assertArrayEquals(expectedResponseStream.toByteArray(), actualResponseStream2.toByteArray());
+
+        // requests must contain our message
+        var req2 = logSink.getOutbound().get(0);
+        var resp2 = logSink.getInbound().get(0);
+
+        assertArrayEquals(actualRequestStream2.toByteArray(), req2.toByteArray());
+        assertArrayEquals(expectedResponseStream.toByteArray(), resp2.toByteArray());
+
+        // ensure request headers are logged
+        assertTrue(
+                logSink.getOutboundHeaders().get(0)
+                        .get(ClientTransportBinding.USER_AGENT_KEY.toLowerCase())
+                        .contains(ClientTransportBinding.USER_AGENT_VALUE)
+        );
+        // ensure response headers are logged
+        assertTrue(
+                logSink.getInboundHeaders().get(0)
+                        .get(HttpServerUtil.GzipResponseHandler.TEST_HEADER_KEY.toLowerCase())
+                        .contains(HttpServerUtil.GzipResponseHandler.TEST_HEADER_VALUE)
+        );
+
+        // all headers must've been converted to lower case, these must be false
+        assertFalse(
+                logSink.getOutboundHeaders().get(0)
+                        .get(ClientTransportBinding.USER_AGENT_KEY)
+                        .contains(ClientTransportBinding.USER_AGENT_VALUE)
+        );
+        assertFalse(
+                logSink.getInboundHeaders().get(0)
+                        .get(HttpServerUtil.GzipResponseHandler.TEST_HEADER_KEY)
+                        .contains(HttpServerUtil.GzipResponseHandler.TEST_HEADER_VALUE)
+        );
+    }
+
+    private void compareTransactionIds(ArrayList<String> inboundTransactionIds, ArrayList<String> outboundTransactionIds) {
+        final var inboundSet = new HashSet<>(inboundTransactionIds);
+        final var outboundSet = new HashSet<>(outboundTransactionIds);
+
+        //check uniqueness of transaction ids
+        assertEquals(inboundTransactionIds.size(), inboundSet.size(), "Duplicate Inbound TransactionIds.");
+        assertEquals(outboundTransactionIds.size(), outboundSet.size(), "Duplicate Outbound TransactionIds.");
+
+        //every request has a response
+        assertEquals(inboundSet, outboundSet, "Not all responses are associated with a request.");
+    }
 
     static class TestCommLogSink implements CommunicationLogSink {
 
@@ -352,6 +506,10 @@ public class CommunicationLogIT extends DpwsTest {
         private final ArrayList<ListMultimap<String, String>> outboundHeaders;
         private final ArrayList<Map<String, String>> inboundHeadersOld;
         private final ArrayList<Map<String, String>> outboundHeadersOld;
+        private CommunicationLog.MessageType inboundMessageType;
+        private CommunicationLog.MessageType outboundMessageType;
+        private final ArrayList<String> inboundTransactionIds;
+        private final ArrayList<String> outboundTransactionIds;
 
         TestCommLogSink() {
             this.inbound = new ArrayList<>();
@@ -360,11 +518,14 @@ public class CommunicationLogIT extends DpwsTest {
             this.outboundHeaders = new ArrayList<>();
             this.inboundHeadersOld = new ArrayList<>();
             this.outboundHeadersOld = new ArrayList<>();
+            this.inboundTransactionIds = new ArrayList<>();
+            this.outboundTransactionIds = new ArrayList<>();
         }
 
         @Override
         public OutputStream createTargetStream(CommunicationLog.TransportType path,
                                                CommunicationLog.Direction direction,
+                                               CommunicationLog.MessageType messageType,
                                                CommunicationContext communicationContext) {
             var os = new ByteArrayOutputStream();
             var appInfo = (HttpApplicationInfo) communicationContext.getApplicationInfo();
@@ -372,10 +533,16 @@ public class CommunicationLogIT extends DpwsTest {
                 inbound.add(os);
                 inboundHeaders.add(appInfo.getHeaders());
                 inboundHeadersOld.add(appInfo.getHttpHeaders());
+                inboundMessageType = messageType;
+                inboundTransactionIds.add(
+                        ((HttpApplicationInfo) communicationContext.getApplicationInfo()).getTransactionId());
             } else {
                 outbound.add(os);
                 outboundHeaders.add(appInfo.getHeaders());
                 outboundHeadersOld.add(appInfo.getHttpHeaders());
+                outboundMessageType = messageType;
+                outboundTransactionIds.add(
+                        ((HttpApplicationInfo) communicationContext.getApplicationInfo()).getTransactionId());
             }
             return os;
         }
@@ -411,6 +578,22 @@ public class CommunicationLogIT extends DpwsTest {
             outboundHeaders.clear();
             inboundHeadersOld.clear();
             outboundHeadersOld.clear();
+        }
+
+        public CommunicationLog.MessageType getOutboundMessageType() {
+            return outboundMessageType;
+        }
+
+        public CommunicationLog.MessageType getInboundMessageType() {
+            return inboundMessageType;
+        }
+
+        public ArrayList<String> getInboundTransactionIds() {
+            return inboundTransactionIds;
+        }
+
+        public ArrayList<String> getOutboundTransactionIds() {
+            return outboundTransactionIds;
         }
     }